Move hand-written test to "tests" folder
Now, "test" only contains generated stubs and all hand-written tests are in "tests".
